// src/app/core/api/model/availableVersionsResponseDto.ts
/**
 * Flui.cloud API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AvailableVersionDto } from './availableVersionDto';


export interface AvailableVersionsResponseDto { 
    sourceType: AvailableVersionsResponseDto.SourceTypeEnum;
    /**
     * Currently deployed image reference
     */
    currentImageRef?: string;
    versions: Array<AvailableVersionDto>;
    /**
     * Next page number for pagination (DockerHub only)
     */
    nextPage?: number | null;
    /**
     * For system apps: glob patterns of versions allowed by Flui curation (e.g. [\"v4.*\"]). null for user apps (no restriction).
     */
    allowedPatterns?: Array<string> | null;
}
export namespace AvailableVersionsResponseDto {
    export const SourceTypeEnum = {
        DockerImage: 'docker_image',
        GitBuild: 'git_build',
        HelmChart: 'helm_chart',
        RawManifest: 'raw_manifest'
    } as const;
    export type SourceTypeEnum = typeof SourceTypeEnum[keyof typeof SourceTypeEnum];
}